dtype

+++ flamedisx/core.py
@@ -328,7 +328,7 @@ class ERSource:
 
         # Precompute tensors for use in gimme
         for fname, v in self.f_dims.items():
-            self.tensor_data[fname] = [tf.convert_to_tensor(d[x]) for x in v]
+            self.tensor_data[fname] = [tf.convert_to_tensor(d[x], dtype=tf.float32) for x in v]
         for fname in ['s1', 's2']:
             self.tensor_data[fname] = tf.convert_to_tensor(d[fname], dtype=tf.float32)
